This commit removes all global dependencies code ( log4js, mongoose, bluebird and more). Reducing the bundle from 4 Mbs to 600 kbs.
Also this commit isolates the log4js dependency with a factory class, the purpose is to use the library for both browser and node.

This commit is a middle step. The bundle files does not work and the scrips for user generation were commented in order to reduce the bundle file size.